requirements.txt
PySide6>=6.6.0
pyusb>=1.2.1
python-rtmidi>=1.5.8
hidapi>=0.14.0  # Optional but good to have
